Info subprocess
Imagemagick remains the best image library I know... and is only usable in C as far as I can figure. So I made a C thing for doing what I need for the python. It's not an extension because ImageMagick leaks memory like you wouldn't believe, and by isolating that to a subprocess I don't have to kill my server every hour or so.

also some general maintenance, sql error in comics